+++ client/test/karma/reader/Pdf-test.js
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import { getInitialAnnotationIconPageCoords } from '../../../app/components/Pdf';
+import { getInitialAnnotationIconPageCoords } from '../../../app/reader/Pdf';
 import { pageCoordsOfRootCoords } from '../../../app/reader/utils';
 
 describe('Pdf', () => {
